[consensus] Rewrite comments (#219)
