+++ consensus/README.md
@@ -2,8 +2,8 @@
 
 [![Crates.io](https://img.shields.io/crates/v/commonware-consensus.svg)](https://crates.io/crates/commonware-consensus)
 
-TBD
+Order opaque messages in a Byzantine environment.
 
-## Status 
+## Status
 
 `commonware-consensus` is **ALPHA** software and is not yet recommended for production use. Developers should expect breaking changes and occasional instability.
