+++ consensus/build.rs
@@ -1,6 +1,6 @@
 use std::io::Result;
 fn main() -> Result<()> {
-    // Proto compilation rules for `fixed` dialect
+    // Proto compilation rules for `simplex` dialect
     let mut config = prost_build::Config::new();
     config.bytes([
         "Signature.public_key",
@@ -8,6 +8,6 @@ fn main() -> Result<()> {
         "Parent.digest",
         "Proposal.payload",
     ]);
-    config.compile_protos(&["src/authority/wire.proto"], &["src/authority/"])?;
+    config.compile_protos(&["src/simplex/wire.proto"], &["src/simplex/"])?;
     Ok(())
 }
