+++ consensus/src/simplex/actors/voter/ingress.rs
@@ -1,4 +1,4 @@
-use crate::authority::wire;
+use crate::simplex::wire;
 use futures::{channel::mpsc, SinkExt};
 
 // If either of these requests fails, it will not send a reply.
