+++ consensus/src/simplex/encoder.rs
