+++ consensus/src/simplex/engine.rs
@@ -13,6 +13,7 @@ use governor::clock::Clock as GClock;
 use rand::{CryptoRng, Rng};
 use tracing::debug;
 
+/// Instance of `simplex` consensus engine.
 pub struct Engine<
     B: Blob,
     E: Clock + GClock + Rng + CryptoRng + Spawner + Storage<B>,
@@ -42,6 +43,7 @@ impl<
         S: Supervisor<Seed = (), Index = View>,
     > Engine<B, E, C, H, A, R, F, S>
 {
+    /// Create a new `simplex` consensus engine.
     pub fn new(runtime: E, journal: Journal<B, E>, cfg: Config<C, H, A, R, F, S>) -> Self {
         // Ensure configuration is valid
         cfg.assert();
@@ -97,6 +99,9 @@ impl<
         }
     }
 
+    /// Start the `simplex` consensus engine.
+    ///
+    /// This will also rebuild the state of the engine from provided `Journal`.
     pub async fn run(
         self,
         voter_network: (impl Sender, impl Receiver),
